Irda: Add Irda System Service for Samsung devices
Add Irda System Service that is compatible with the Samsung devices
.
This is not a perfect solution yet, but it does work and should therefore be a starting point.
I do not know how to make this code exclusive to the S4, maybe someone can help me out with that.

PatchSet 2: Now the Irda service is not started by default. It can be enabled via overlay.

PatchSet 3: Rebase

PatchSet 4: It now uses a HAL module to communicate with sysfs. However it currently
fails when loading the module

PatchSet 5: Corrected a typo in a comment

PatchSet 6: Cleanup

PatchSet 7: Renamed native method to send_ircode_native

Patchset 8: Renamed commit message to indicate general applicability to Samsung devices
including n80xx.  Need to determine how to handle HTC and Sony...  Also added logging
when module load fails or device opening fails.

PatchSet 9: Fix missing semicolon causing breakage in the build

Change-Id: Id4da55cfbd297fa46c50aa8a6af9cd4d056086cd

Conflicts:
	core/res/res/values/config.xml
	core/res/res/values/symbols.xml
	services/java/com/android/server/SystemServer.java